+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Split show subcommand:
     - Export 
     - Show
+- Change option to the generate-password subcommand
 
 ## [0.8.0]
 - [cithare 0.8.0](https://github.com/EruEri/cithare)